Don't send 0 for numeric parmeters that have not been specified (or false for boolean).
Parameters should be boxed (object type) in internal storage, so if the user did not specify the parameter,
we will not assume the value is 0 - as in ARI not specifying an optional param often has more meaning than
"zero". For example, in `POST /channels` ("channel originate"), not setting the priority means "1" and not "0".